+++ go/src/github.com/bitrise-core/bitrise-init/cli/init_manual_config.go
@@ -20,14 +20,14 @@ import (
 	envmanModels "github.com/bitrise-io/envman/models"
 	"github.com/bitrise-io/go-utils/colorstring"
 	"github.com/bitrise-io/go-utils/pathutil"
-	"github.com/codegangsta/cli"
+	"github.com/urfave/cli"
 )
 
 const (
 	defaultOutputDir = "_defaults"
 )
 
-func initManualConfig(c *cli.Context) {
+func initManualConfig(c *cli.Context) error {
 	PrintHeader(c)
 
 	//
@@ -38,7 +38,7 @@ func initManualConfig(c *cli.Context) {
 
 	currentDir, err := pathutil.AbsPath("./")
 	if err != nil {
-		log.Fatalf("Failed to get current directory, error: %s", err)
+		return fmt.Errorf("Failed to get current directory, error: %s", err)
 	}
 
 	if outputDir == "" {
@@ -46,7 +46,7 @@ func initManualConfig(c *cli.Context) {
 	}
 	outputDir, err = pathutil.AbsPath(outputDir)
 	if err != nil {
-		log.Fatalf("Failed to get abs path (%s), error: %s", outputDir, err)
+		return fmt.Errorf("Failed to get abs path (%s), error: %s", outputDir, err)
 	}
 
 	if formatStr == "" {
@@ -54,10 +54,10 @@ func initManualConfig(c *cli.Context) {
 	}
 	format, err := output.ParseFormat(formatStr)
 	if err != nil {
-		log.Fatalf("Failed to parse format, err: %s", err)
+		return fmt.Errorf("Failed to parse format, err: %s", err)
 	}
 	if format != output.JSONFormat && format != output.YAMLFormat {
-		log.Fatalf("Not allowed output format (%v), options: [%s, %s]", format, output.YAMLFormat.String(), output.JSONFormat.String())
+		return fmt.Errorf("Not allowed output format (%v), options: [%s, %s]", format, output.YAMLFormat.String(), output.JSONFormat.String())
 	}
 
 	if isCI {
@@ -88,7 +88,7 @@ func initManualConfig(c *cli.Context) {
 		log.Debug("Analyze result:")
 		bytes, err := yaml.Marshal(option)
 		if err != nil {
-			log.Fatalf("Failed to marshal option, err: %s", err)
+			return fmt.Errorf("Failed to marshal option, err: %s", err)
 		}
 		log.Debugf("\n%v", string(bytes))
 
@@ -96,7 +96,7 @@ func initManualConfig(c *cli.Context) {
 
 		configs, err := detector.DefaultConfigs()
 		if err != nil {
-			log.Fatalf("Failed create default configs, error: %s", err)
+			return fmt.Errorf("Failed create default configs, error: %s", err)
 		}
 
 		for name, config := range configs {
@@ -104,7 +104,7 @@ func initManualConfig(c *cli.Context) {
 
 			bytes, err := yaml.Marshal(config)
 			if err != nil {
-				log.Fatalf("Failed to marshal option, err: %s", err)
+				return fmt.Errorf("Failed to marshal option, err: %s", err)
 			}
 			log.Debugf("\n%v", string(bytes))
 		}
@@ -114,7 +114,7 @@ func initManualConfig(c *cli.Context) {
 
 	customConfigs, err := scanners.CustomConfig()
 	if err != nil {
-		log.Fatalf("Failed create default custom configs, error: %s", err)
+		return fmt.Errorf("Failed create default custom configs, error: %s", err)
 	}
 
 	projectTypeConfigMap["custom"] = customConfigs
@@ -130,39 +130,16 @@ func initManualConfig(c *cli.Context) {
 		}
 
 		if err := os.MkdirAll(outputDir, 0700); err != nil {
-			log.Fatalf("Failed to create (%s), err: %s", outputDir, err)
+			return fmt.Errorf("Failed to create (%s), err: %s", outputDir, err)
 		}
 
 		pth := path.Join(outputDir, "result")
 		if err := output.Print(scanResult, format, pth); err != nil {
-			log.Fatalf("Failed to print result, error: %s", err)
+			return fmt.Errorf("Failed to print result, error: %s", err)
 		}
 		log.Infof("  scan result: %s", colorstring.Blue(pth))
 
-		return
-	}
-
-	//
-	// Write output to files
-	if isCI {
-		log.Infof(colorstring.Blue("Saving outputs:"))
-
-		scanResult := models.ScanResultModel{
-			OptionsMap: projectTypeOptionMap,
-			ConfigsMap: projectTypeConfigMap,
-		}
-
-		if err := os.MkdirAll(outputDir, 0700); err != nil {
-			log.Fatalf("Failed to create (%s), error: %s", outputDir, err)
-		}
-
-		pth := path.Join(outputDir, "result")
-		if err := output.Print(scanResult, format, pth); err != nil {
-			log.Fatalf("Failed to print result, error: %s", err)
-		}
-		log.Infof("  scan result: %s", colorstring.Blue(pth))
-
-		return
+		return nil
 	}
 
 	//
@@ -175,27 +152,27 @@ func initManualConfig(c *cli.Context) {
 		// Init
 		platformOutputDir := path.Join(outputDir, detectorName)
 		if exist, err := pathutil.IsDirExists(platformOutputDir); err != nil {
-			log.Fatalf("Failed to check if path (%s) exis, error: %s", platformOutputDir, err)
+			return fmt.Errorf("Failed to check if path (%s) exis, error: %s", platformOutputDir, err)
 		} else if exist {
 			if err := os.RemoveAll(platformOutputDir); err != nil {
-				log.Fatalf("Failed to cleanup (%s), error: %s", platformOutputDir, err)
+				return fmt.Errorf("Failed to cleanup (%s), error: %s", platformOutputDir, err)
 			}
 		}
 
 		if err := os.MkdirAll(platformOutputDir, 0700); err != nil {
-			log.Fatalf("Failed to create (%s), error: %s", platformOutputDir, err)
+			return fmt.Errorf("Failed to create (%s), error: %s", platformOutputDir, err)
 		}
 
 		// Collect inputs
 		configPth := ""
 		appEnvs := []envmanModels.EnvironmentItemModel{}
 
-		var walkDepth func(option models.OptionModel)
+		var walkDepth func(option models.OptionModel) error
 
-		walkDepth = func(option models.OptionModel) {
+		walkDepth = func(option models.OptionModel) error {
 			optionEnvKey, selectedValue, err := askForValue(option)
 			if err != nil {
-				log.Fatalf("Failed to ask for vale, error: %s", err)
+				return fmt.Errorf("Failed to ask for vale, error: %s", err)
 			}
 
 			if optionEnvKey == "" {
@@ -208,19 +185,21 @@ func initManualConfig(c *cli.Context) {
 
 			nestedOptions, found := option.ValueMap[selectedValue]
 			if !found {
-				return
+				return err
 			}
 
-			walkDepth(nestedOptions)
+			return walkDepth(nestedOptions)
 		}
 
-		walkDepth(option)
+		if err := walkDepth(option); err != nil {
+			return err
+		}
 
 		log.Debug()
 		log.Debug("Selected app envs:")
 		aBytes, err := yaml.Marshal(appEnvs)
 		if err != nil {
-			log.Fatalf("Failed to marshal appEnvs, error: %s", err)
+			return fmt.Errorf("Failed to marshal appEnvs, error: %s", err)
 		}
 		log.Debugf("\n%v", string(aBytes))
 
@@ -229,7 +208,7 @@ func initManualConfig(c *cli.Context) {
 
 		var config bitriseModels.BitriseDataModel
 		if err := yaml.Unmarshal([]byte(configStr), &config); err != nil {
-			log.Fatalf("Failed to unmarshal config, error: %s", err)
+			return fmt.Errorf("Failed to unmarshal config, error: %s", err)
 		}
 
 		config.App.Environments = appEnvs
@@ -239,16 +218,18 @@ func initManualConfig(c *cli.Context) {
 		log.Debugf("  name: %s", configPth)
 		aBytes, err = yaml.Marshal(config)
 		if err != nil {
-			log.Fatalf("Failed to marshal config, error: %s", err)
+			return fmt.Errorf("Failed to marshal config, error: %s", err)
 		}
 		log.Debugf("\n%v", string(aBytes))
 
 		// Write config to file
 		pth := path.Join(platformOutputDir, configPth)
 		if err := output.Print(config, format, pth); err != nil {
-			log.Fatalf("Failed to print result, error: %s", err)
+			return fmt.Errorf("Failed to print result, error: %s", err)
 		}
 		log.Infof("  bitrise.yml template: %s", colorstring.Blue(pth))
 		fmt.Println()
 	}
+
+	return nil
 }